Move permission denied popover to capture receipt component
