+++ src/app/shared/components/capture-receipt/camera-preview/camera-preview.component.ts
@@ -2,12 +2,8 @@ import { Component, Input, OnInit, Output, EventEmitter, OnChanges, SimpleChange
 import { CameraPreview, CameraPreviewOptions } from '@capacitor-community/camera-preview';
 import { Camera } from '@capacitor/camera';
 import { Capacitor } from '@capacitor/core';
-import { Platform, PopoverController } from '@ionic/angular';
-import { AndroidSettings, IOSSettings, NativeSettings } from 'capacitor-native-settings';
 import { from } from 'rxjs';
-import { tap, switchMap } from 'rxjs/operators';
 import { LoaderService } from 'src/app/core/services/loader.service';
-import { PopupAlertComponentComponent } from '../../popup-alert-component/popup-alert-component.component';
 
 @Component({
   selector: 'app-camera-preview',
@@ -39,6 +35,8 @@ export class CameraPreviewComponent implements OnInit, OnChanges {
 
   @Output() dismissCameraPreview = new EventEmitter();
 
+  @Output() permissionDenied = new EventEmitter<'CAMERA' | 'GALLERY'>();
+
   isCameraPreviewInitiated = false;
 
   isCameraPreviewStarted = false;
@@ -47,11 +45,7 @@ export class CameraPreviewComponent implements OnInit, OnChanges {
 
   showModeChangedMessage = false;
 
-  constructor(
-    private loaderService: LoaderService,
-    private popoverController: PopoverController,
-    private platform: Platform
-  ) {}
+  constructor(private loaderService: LoaderService) {}
 
   setUpAndStartCamera() {
     if (Capacitor.getPlatform() === 'web') {
@@ -61,7 +55,7 @@ export class CameraPreviewComponent implements OnInit, OnChanges {
 
     from(Camera.requestPermissions()).subscribe((permissions) => {
       if (permissions?.camera === 'denied') {
-        return this.showPermissionDeniedPopover('CAMERA');
+        return this.permissionDenied.emit('CAMERA');
       }
 
       /*
@@ -161,55 +155,5 @@ export class CameraPreviewComponent implements OnInit, OnChanges {
         this.showModeChangedMessage = false;
       }, 1000);
     }
-  }
-
-  setupPermissionDeniedPopover(permissionType: 'CAMERA' | 'GALLERY') {
-    const isIos = this.platform.is('ios');
-
-    const galleryPermissionName = isIos ? 'Photos' : 'Storage';
-    let title = 'Camera Permission';
-    if (permissionType === 'GALLERY') {
-      title = galleryPermissionName + ' Permission';
-    }
-
-    const cameraPermissionMessage = `To capture photos, please allow Fyle to access your camera. Click Settings and allow access to Camera and ${galleryPermissionName}`;
-    const galleryPermissionMessage = `Please allow Fyle to access device photos. Click Settings and allow ${galleryPermissionName} access`;
-
-    const message = permissionType === 'CAMERA' ? cameraPermissionMessage : galleryPermissionMessage;
-
-    return this.popoverController.create({
-      component: PopupAlertComponentComponent,
-      componentProps: {
-        title,
-        message,
-        primaryCta: {
-          text: 'Open Settings',
-          action: 'OPEN_SETTINGS',
-        },
-        secondaryCta: {
-          text: 'Cancel',
-          action: 'CANCEL',
-        },
-      },
-      cssClass: 'pop-up-in-center',
-      backdropDismiss: false,
-    });
-  }
-
-  showPermissionDeniedPopover(permissionType: 'CAMERA' | 'GALLERY') {
-    from(this.setupPermissionDeniedPopover(permissionType))
-      .pipe(
-        tap((permissionDeniedPopover) => permissionDeniedPopover.present()),
-        switchMap((permissionDeniedPopover) => permissionDeniedPopover.onWillDismiss())
-      )
-      .subscribe(({ data }) => {
-        if (data?.action === 'OPEN_SETTINGS') {
-          NativeSettings.open({
-            optionAndroid: AndroidSettings.ApplicationDetails,
-            optionIOS: IOSSettings.App,
-          });
-        }
-        this.onDismissCameraPreview();
-      });
   }
 }
