+++ src/app/shared/components/capture-receipt/capture-receipt.component.html
@@ -12,4 +12,5 @@
   (receiptPreview)="openReceiptPreviewModal()"
   (dismissCameraPreview)="onDismissCameraPreview()"
   (captureReceipt)="onCaptureReceipt()"
+  (permissionDenied)="showPermissionDeniedPopover($event)"
 ></app-camera-preview>
